requirements.txt
Flask==2.3.3
Pillow==10.0.1
reportlab==4.0.4
pikepdf==8.4.1